refactor: split out screenshot specific functions
