+++ utils.py
@@ -2,7 +2,7 @@ from selenium import webdriver
 from selenium.webdriver.chrome.service import Service
 from webdriver_manager.chrome import ChromeDriverManager
 from pathlib import Path
-import os
+import sys, os
 
 from config import SCREEN_SIZES, FOLDER_DIRECTORIES
 
@@ -34,3 +34,30 @@ def create_webdriver(device_type):
 
 def get_folder(device_type):
     return Path(FOLDER_DIRECTORIES[device_type])
+
+def open_screenshots(folder):
+    if sys.platform == "win32":
+        os.startfile(folder)
+    elif sys.platform == "darwin":
+        os.system('open "%s"' % folder)
+    elif sys.platform == "linux":
+        os.system('xdg-open "%s"' % folder)
+
+def rename_file(filename):
+    # files have naming restrictions so saving file as the name of the URL (the below list covers filename forbidden
+    # characters). Not all below are valid URL characters but if ever functionality changed from URL to something else
+    # as the filename, the below will cover all restrictions. The file name can be changed below or additional
+    # restrictions handled
+    filename = filename.replace("://", "_")
+    filename = filename.replace("\\", "_")
+    filename = filename.replace("/", "_")
+    filename = filename.replace(":", "_")
+    filename = filename.replace("*", "_")
+    filename = filename.replace("?", "_")
+    filename = filename.replace("\"", "_")
+    filename = filename.replace("<", "_")
+    filename = filename.replace(">", "_")
+    filename = filename.replace("|", "_")
+    filename = filename.replace(".", "_")
+
+    return filename
